scope dates utils
