+++ components/widgets/forest-change/glads/selectors.js
@@ -3,21 +3,21 @@ import moment from 'moment';
 import { format } from 'd3-format';
 import isEmpty from 'lodash/isEmpty';
 import sumBy from 'lodash/sumBy';
-import { buildDateArray } from 'utils/dates';
+import { buildDateArray } from 'components/widgets/utils/dates';
 
 // get list data
-const getData = state => state.data && state.data.alerts;
-const getLatestDate = state => state.data && state.data.latestDate;
-const getCurrentLocation = state => state.locationLabel;
-const getColors = state => state.colors;
-const getSentence = state => state.sentence;
+const getData = (state) => state.data && state.data.alerts;
+const getLatestDate = (state) => state.data && state.data.latestDate;
+const getCurrentLocation = (state) => state.locationLabel;
+const getColors = (state) => state.colors;
+const getSentence = (state) => state.sentence;
 
 export const filterData = createSelector(
   [getData, getLatestDate],
   (data, latestDate) => {
     if (!data || isEmpty(data)) return null;
 
-    const filteredData = data.filter(d =>
+    const filteredData = data.filter((d) =>
       moment(d.day).isAfter(moment(latestDate).subtract(1, 'week'))
     );
 
@@ -26,12 +26,12 @@ export const filterData = createSelector(
     const startDate = moment(latestDate).subtract(1, 'week');
     const endDate = moment(latestDate);
 
-    const weekData = buildDateArray(startDate, endDate).map(d => {
-      const dayData = filteredData.find(fd => fd.day === d);
+    const weekData = buildDateArray(startDate, endDate).map((d) => {
+      const dayData = filteredData.find((fd) => fd.day === d);
       return (
         dayData || {
           count: 0,
-          day: d
+          day: d,
         }
       );
     });
@@ -40,7 +40,7 @@ export const filterData = createSelector(
   }
 );
 
-export const parseData = createSelector([filterData], data => {
+export const parseData = createSelector([filterData], (data) => {
   if (!data || data.length < 1) return null;
   return data;
 });
@@ -58,48 +58,48 @@ export const parseConfig = createSelector(
             x1: '0%',
             y1: '0%',
             x2: '100%',
-            y2: '0%'
+            y2: '0%',
           },
           stops: [
             {
               offset: '0%',
               stopColor: colors.main,
-              stopOpacity: '0'
+              stopOpacity: '0',
             },
             {
               offset: '100%',
-              stopColor: colors.main
-            }
-          ]
-        }
+              stopColor: colors.main,
+            },
+          ],
+        },
       ],
       xKey: 'day',
       yKeys: {
         lines: {
           count: {
             stroke: 'url(#firesGradient)',
-            type: 'monotone'
-          }
-        }
+            type: 'monotone',
+          },
+        },
       },
       xAxis: {
         tickCount: 2,
         interval: data.length - 2,
         padding: { right: 8 },
-        tickFormatter: t => moment(t).format('Do MMM')
+        tickFormatter: (t) => moment(t).format('Do MMM'),
       },
       tooltip: [
         {
           key: 'day',
-          unitFormat: value => moment(value).format('Do MMM YYYY')
+          unitFormat: (value) => moment(value).format('Do MMM YYYY'),
         },
         {
           key: 'count',
           unit: '',
-          unitFormat: value =>
-            (Number.isInteger(value) ? format(',')(value) : value)
-        }
-      ]
+          unitFormat: (value) =>
+            Number.isInteger(value) ? format(',')(value) : value,
+        },
+      ],
     };
   }
 );
@@ -117,7 +117,7 @@ export const parseSentence = createSelector(
         typeof weekGladCount === 'number'
           ? format(',')(weekGladCount)
           : weekGladCount,
-      weeklyMean: format(',')(averageGladCount)
+      weeklyMean: format(',')(averageGladCount),
     };
     return { sentence, params };
   }
@@ -126,5 +126,5 @@ export const parseSentence = createSelector(
 export default createStructuredSelector({
   data: parseData,
   config: parseConfig,
-  sentence: parseSentence
+  sentence: parseSentence,
 });
